trap signals only on posix systems.
