+++ trap_nonposix.go
@@ -1,0 +1,8 @@
+// +build windows plan9 nacl js
+
+package p2pd
+
+func (d *Daemon) trapSignals() {
+	// TODO: define signals we want to trap on Windows, if any.
+	return
+}
